YTDB-558: Route VertexStep recognition to dedicated VertexHopRecogniser/EdgeHopRecogniser
TinkerPop models both a vertex-returning hop (out/in/both, plus the
folded outE().inV() chains) and an edge-returning step (outE/inE/bothE)
with a single VertexStep class, distinguished only by returnsEdge().
The old VertexStepRecogniser conflated two responsibilities: it handled
the bare-hop branch itself and delegated the edge branch to the
edge-filter recogniser, so the bare-hop logic and the routing decision
lived in one class and could not be tested or reasoned about
independently.

Split the concerns behaviour-preservingly:

- VertexHopRecogniser (new) owns the bare-hop logic verbatim, with its
  own instanceof guard so a direct mis-call declines with 0 rather than
  throwing (contract parity with the edge recogniser).
- EdgeStepRecogniser is renamed to EdgeHopRecogniser for symmetry; its
  recognition logic and precondition guard are unchanged.
- VertexStepRecogniser becomes a thin router registered under
  VertexStep.class: it re-asserts the step type, then forwards to
  EdgeHopRecogniser or VertexHopRecogniser on returnsEdge(), returning
  the handler's int result verbatim. No behaviour change for any query.

Tests are repartitioned to match: the moved bare-hop cases now target
VertexHopRecogniser directly (plus a non-VertexStep decline test for its
guard), the renamed EdgeHopRecogniserTest keeps its coverage, and
VertexStepRecogniserTest is reframed as a router test that pins both
returnsEdge() arms and the guard, closing the router's branch-coverage
gap. Stale {@link}s left by the rename are repointed, with references to
the specific bare-hop behaviour aimed at VertexHopRecogniser and
references to the class that owns/registers VertexStep.class left on the
router.
